Moved and renamed funcs in bootstrap.c/policy_server.c
